Change Power.toArray() return type to double

+++ src/main/java/name/mjw/jquante/math/qm/basis/Power.java
@@ -99,8 +99,8 @@ public class Power {
 	 *
 	 * @return powers
 	 */
-	public int[] toArray() {
-		return new int[] { l, m, n };
+	public double[] toArray() {
+		return new double[] { l, m, n };
 	}
 
 	/**
